Вынес классы отдельно
